pyproject.toml
[project]
name = "stamping-inspection-cameras-ctrl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.2",
    "aio-pika>=9.5.7",
    "apscheduler>=3.11.0",
    "asyncodbc>=0.1.1",
    "dbutils>=3.1.2",
    "munpy>=0.0.0",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "psutil>=7.1.0",
    "pymodbus>=3.11.3",
    "pymssql>=2.3.8",
    "pymysql>=1.1.2",
    "pyodbc>=5.3.0",
    "python-snap7>=2.0.2",
    "pyyaml>=6.0.3",
    "redis>=6.4.0",
    "tortoise-orm>=0.25.1",
]

[tool.aerich]
tortoise_orm = "config.mssql_setting.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
